chore(http): Directory restructuring
- Moved RequestHandlers to their own directory
