+++ http/tester/main.cpp
@@ -1,5 +1,7 @@
+#include "GetRequestHandler.hpp"
 #include "IRequestHandler.hpp"
 #include "Request.hpp"
+#include "RequestHandlerFactory.hpp"
 #include "Response.hpp"
 #include "webserv.hpp"
 #include <iostream>
